feat : Car Class 생성 및 TestCode 구현
